+++ src/hump_yard/__init__.py
@@ -1,7 +1,9 @@
-# hump_yard package
+"""
+Hump Yard - File monitoring daemon with plugin support.
+"""
 from .daemon import FileMonitorDaemon
 from .base_plugin import FileProcessorPlugin, PluginManager
 
 __all__ = ['FileMonitorDaemon', 'FileProcessorPlugin', 'PluginManager']
 
-__version__ = '0.2.0'
+__version__ = '0.3.0'