+++ src/hump_yard/plugins/__init__.py
@@ -1,1 +1,3 @@
-# Plugins package
+"""
+Plugins package for hump-yard file processor plugins.
+"""